# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: NEXUS Super Social Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /challenges:
    get:
      operationId: listChallenges
      tags: [challenges]
      summary: List active challenges
      parameters:
        - in: query
          name: creatorId
          schema: { type: integer }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Challenge"
    post:
      operationId: createChallenge
      tags: [challenges]
      summary: Create a new video challenge (creator pays prize pool from wallet)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChallengeInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Challenge"
        "401":
          description: Unauthorized
        "400":
          description: Insufficient wallet balance for prize pool
  /challenges/{id}:
    get:
      operationId: getChallenge
      tags: [challenges]
      summary: Get a single challenge with participant count
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Challenge"
        "404":
          description: Not found
  /challenges/{id}/join:
    post:
      operationId: joinChallenge
      tags: [challenges]
      summary: Join a challenge with a reel submission
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/JoinChallengeInput"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Challenge"
        "401":
          description: Unauthorized

  # ── GAMIFICATION (streaks) ───────────────────────────────────────────────
...
components:
  schemas:
...
    User:
      type: object
      required: [id, username, displayName, email, createdAt]
      properties:
        id: { type: integer }
        username: { type: string }
        displayName: { type: string }
        email: { type: string }
        bio: { type: ["string", "null"] }
        avatarUrl: { type: ["string", "null"] }
        coverUrl: { type: ["string", "null"] }
        followersCount: { type: integer }
        followingCount: { type: integer }
        postsCount: { type: integer }
        isVerified: { type: boolean }
        isFollowing: { type: boolean }
        status: { type: string, enum: [active, suspended] }
        createdAt: { type: string }
...
    ChallengeInput:
      type: object
      required: [name, hashtag, category, days, prizePool, judgeType]
      properties:
        name: { type: string }
        hashtag: { type: string }
        category: { type: string }
        description: { type: string }
        days: { type: integer, default: 7 }
        prizePool: { type: integer, default: 0 }
        judgeType: { type: string, enum: [vote, ai, views], default: vote }
        settings: { type: object, additionalProperties: true }
    Challenge:
      type: object
      required: [id, creatorId, name, hashtag, category, days, prizePool, judgeType, status, startsAt, createdAt, participantsCount]
      properties:
        id: { type: integer }
        creatorId: { type: integer }
        creator:
          $ref: "#/components/schemas/User"
        name: { type: string }
        hashtag: { type: string }
        category: { type: string }
        description: { type: ["string", "null"] }
        days: { type: integer }
        prizePool: { type: integer }
        judgeType: { type: string }
        status: { type: string, enum: [active, ended] }
        settings: { type: object, additionalProperties: true }
        startsAt: { type: string }
        endsAt: { type: ["string", "null"] }
        createdAt: { type: string }
        participantsCount: { type: integer }
        isJoined: { type: boolean }
    JoinChallengeInput:
      type: object
      required: [reelId]
      properties:
        reelId: { type: integer }

    # ── Gamification ────────────────────────────────────────────────────────
...